+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update minimum supported buntu version to 20.04+
 - Add support for Czech Crown (CZK)
 - Add support for Zloty (PLN)
+- Add advanced setup with skip microSD card or 12-word seed options
 
 ## 4.37.0
 - Bundle BitBox02 firmware version v9.14.0